ECFLOW-1787: Technology demo: Token authentication
Commit includes a demonstration of how ecFlow
server could support token-based authentication.

In this example JWT tokens are used, and an
authentication service is expected to be found
from host ecflow-auth:5001.

Tokens incude "additional claims" which in this
case are ecFlow paths that the user can access
in either ro or rw mode.

Parsing json is done with nlohmann/json library.